// main-backend/frontend/src/generated/model/meal-in-menu-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MealStatus } from './meal-status';


export interface MealInMenuDto { 
    /**
     * Menu Item Identifier
     */
    id: number;
    /**
     * Name of the meal
     */
    name: string;
    status: MealStatus;
    /**
     * Number of people that are having this meal
     */
    numberOfPeople: number;
    /**
     * Position of the meal within the menu
     */
    position: number;
}