// main-backend/frontend/src/generated/model/organization-role-enum.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Defines the possible roles of an organization member.
 */
export enum OrganizationRoleEnum {
    Member = 'member',
    Admin = 'admin',
    Owner = 'owner',
    Planner = 'planner',
    Shopper = 'shopper',
    Invited = 'invited'
}