// main-backend/frontend/src/generated/model/snapshot-in-menu-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SnapshotInMenuDto { 
    /**
     * Menu Item Identifier
     */
    id: number;
    /**
     * Name of the snapshot
     */
    name: string;
    /**
     * Number of meals included in the snapshot
     */
    numberOfMealsIncluded: number;
    /**
     * Number of total ingredients across all included meals
     */
    numberOfTotalIngredients: number;
    /**
     * Number of ingredients that are available in the stash
     */
    numberOfAvailableIngredients: number;
    /**
     * Position of the snapshot within the menu
     */
    position: number;
}